added controller refactor code
