+++ tests/test_files/conf_invalid_access_control.yaml
@@ -14,11 +14,11 @@ security:
         validation_checks: all
     cors:
       enabled: True
-access_control:
-    api_specs: foca/access_control/api/access-control-specs.yaml
-    api_controllers: foca/access_control/access_control_server
-    db_name: test_db
-    collection_name: test_collection
-    model: foca/access_control/api/default_model.conf
-    owner_headers: ["owner"]
-    user_headers: ["user"]
+    access_control:
+        api_specs: foca/security/access_control/api/access-control-specs.yaml
+        api_controllers: foca/security/access_control/access_control_server.py
+        db_name: test_db
+        collection_name: test_collection
+        model: foca/security/access_control/api/default_model.conf
+        owner_headers: ["owner"]
+        user_headers: ["user"]
