+++ tests/test_files/conf_valid_access_control.yaml
@@ -1,8 +1,25 @@
-access_control:
-    api_specs: foca/access_control/api/access-control-specs.yaml
-    api_controllers: foca/access_control/access_control_server
-    db_name: test_db
-    collection_name: test_collection
-    model: foca/access_control/api/default_model.conf
-    owner_headers: ["owner"]
-    user_headers: ["user"]
+security:
+    auth:
+        add_key_to_claims: True
+        algorithms:
+          - RS256
+        allow_expired: False
+        audience: null
+        claim_identity: sub
+        claim_issuer: iss
+        validation_methods:
+          - userinfo
+          - public_key
+        validation_checks: all
+    cors:
+      enabled: False
+    cors:
+      enabled: True
+    access_control:
+        api_specs: foca/security/access_control/api/access-control-specs.yaml
+        api_controllers: foca/security/access_control/access_control_server.py
+        db_name: test_db
+        collection_name: test_collection
+        model: foca/security/access_control/api/default_model.conf
+        owner_headers: ["owner"]
+        user_headers: ["user"]
